SCMOD-7698: Fixing workflow-worker poison msg handling.
When the workflow-worker is processing a document, if the document
already has values for the CAF_WORKFLOW_SETTINGS field, it means
that the workflow worker has already resolved the arguments and added
them to the document.

This scenario can occur whenever a downstream worker sends a poison
message back to the workflow worker.

In this case, the ArgumentsManager should not try to re-resolve the arguments
again, but instead just copy the value from CAF_WORKFLOW_SETTINGS from
the document field  into the custom data of the document task response, and
return.

Validation has also been added here to guard against unexpected settings
present in CAF_WORKFLOW_SETTINGS, but not present in the workflow
arguments.

Full description of how to reproduce and examples of poison messages
can be found in the Jira.