SCMOD-7698: Updating poison message checks.
To check if a message/document is poison we'll now
check that:

1. It has a a non-empty CAF_WORKFLOW_SETTINGS field.

2. It does NOT have a 'workflowName' inside it's custom data.

Previously we were checking for the absense of a tenantId inside
the custom data, but there are cases where a tenantId may not be
provided (I don't think we have any concrete cases of this at
the moment, but it is allowable).

Instead we'll now check for the absense of a workflowName inside
custom data, there are currently no scenarios where a valid
document will not have this, so if we don't have it, we have some
more confidence we are dealing with a poison message.

Also removed the validation related to CAF_WORKFLOW_SETTINGS
as we didn't feel it was adding much value.